# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CHEMIA Chemistry Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  username: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  username: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Element {
  atomicNumber: number;
  symbol: string;
  name: string;
  atomicMass: number;
  category: string;
  /** @nullable */
  group?: number | null;
  period: number;
  electronConfiguration: string;
  oxidationStates: string;
  /** @nullable */
  electronegativity?: number | null;
  /** @nullable */
  atomicRadius?: number | null;
  /** @nullable */
  meltingPoint?: number | null;
  /** @nullable */
  boilingPoint?: number | null;
  /** @nullable */
  density?: number | null;
  /** @nullable */
  discoveredBy?: string | null;
  phase: string;
  color: string;
}

export interface ElementCategoryStat {
  category: string;
  count: number;
}

export interface FormulaInput {
  formula: string;
}

export interface ElementComposition {
  symbol: string;
  name: string;
  count: number;
  atomicMass: number;
  totalMass: number;
  percentComposition: number;
}

export interface MolecularMassResult {
  formula: string;
  molecularMass: number;
  molarMass: number;
  composition: ElementComposition[];
}

export interface ParsedElement {
  symbol: string;
  count: number;
}

export interface ParsedFormula {
  formula: string;
  isValid: boolean;
  elements: ParsedElement[];
  totalAtoms: number;
  /** @nullable */
  error?: string | null;
}

export interface EquationInput {
  equation: string;
}

export interface EquationSide {
  formula: string;
  coefficient: number;
}

export interface BalancedEquation {
  original: string;
  balanced: string;
  reactants: EquationSide[];
  products: EquationSide[];
  isBalanced: boolean;
  steps: string[];
}

export interface CompoundBuilderInput {
  element1Symbol: string;
  element1Oxidation: number;
  element2Symbol: string;
  element2Oxidation: number;
}

export interface CompoundElement {
  symbol: string;
  name: string;
  oxidation: number;
  count: number;
}

export interface CompoundResult {
  formula: string;
  name: string;
  element1: CompoundElement;
  element2: CompoundElement;
  molecularMass: number;
  description: string;
}

export interface SavedFormula {
  id: number;
  formula: string;
  name: string;
  type: string;
  /** @nullable */
  molecularMass?: number | null;
  createdAt: string;
}

export interface SavedFormulaInput {
  formula: string;
  name: string;
  type: string;
  /** @nullable */
  molecularMass?: number | null;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: ChatMessage[];
}

export interface MessageInput {
  content: string;
}

export interface CalculationRecord {
  id: number;
  type: string;
  input: string;
  result: string;
  /** @nullable */
  userId?: number | null;
  createdAt: string;
}

export type ListElementsParams = {
search?: string;
category?: string;
};