# lib/api-zod/src/generated/types/compoundBuilderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CHEMIA Chemistry Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CompoundBuilderInput {
  element1Symbol: string;
  element1Oxidation: number;
  element2Symbol: string;
  element2Oxidation: number;
}