# lib/api-zod/src/generated/types/elementComposition.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CHEMIA Chemistry Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ElementComposition {
  symbol: string;
  name: string;
  count: number;
  atomicMass: number;
  totalMass: number;
  percentComposition: number;
}